Commit punto 2
Se genero un codigo para el punto 2, falta configurar el perimetro del cuadrado y el area y perimetro del triangulo

+++ EjemploFiguras(2)/Figura.java
@@ -13,7 +13,7 @@ public abstract class Figura {
     public void setDescripcion(String descripcion) {
         this.descripcion = descripcion;
     }
-     public abstract int calcularArea();
-     public abstract int calcularPerimetro();
+     public abstract double calcularArea();
+     public abstract double calcularPerimetro();
      public abstract void reDibujar();
 }
